Bohubrihi/Rest Api Djanog and React/Django Api/My_Api/requirements.txt
asgiref==3.4.1
Django==3.2.5
django-cleanup==5.2.0
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
Markdown==3.3.4
Pillow==8.3.1
pytz==2021.1
sqlparse==0.4.1
